Reformatted index.html, css changed to offset this
Some of the html in index.html was redundant, i.e. was more or
less 'fluffying' up the code and increasing nesting unnecessarily.
The key change is the hero is now in its own container instead of being
nested in an unnecessary 'head' div.

Obviously, style.css was also reformatted slightly to account for this
change. An offhanded change however is that all children items of
the body now flexes such the footer is at the bottom of the page instead
of floating at the bottom of the contents.

Other than that, no significant change from a client pov.