Finished formatting rand-inf + squashed a bug
The bug was simply that the quote author of the quote would not stick
'relatively' close to the quote it was attached to on a large screen
and would simply move the the rightmost edge of the page (respecting
page padding). This was not expected behaviour, some reorganising
of the html and css fixed this issue however.
- also the year for the 'copyright' was 2021, this has been fixed.

I've also assigned a maximum width to the quote and popup/sign up
so both stick closer to the middle of the page.

At this point this project is more or less finished (according to TOP
specs anyway (linked in README.md)). Any further changes will likely
be irrespective of TOP specifications and of my own free will...
unless we revisit the landing page later in TOP.